Fix build issues
Fix casting issues that are causing the build to fail.

+++ src/chassiscommands.cpp
@@ -512,7 +512,7 @@ ipmi::RspType<bool,    // Power is on
     constexpr bool coolingFanFault = false;
     // chassisIdentifySupport set because this command is implemented
     constexpr bool chassisIdentifySupport = true;
-    uint2_t chassisIdentifyState = chassisIDState;
+    uint2_t chassisIdentifyState = static_cast<uint8_t>(chassisIDState);
     constexpr bool sleepButtonDisabled = false;
     constexpr bool sleepButtonDisableAllow = false;
